requirements.txt
# Compile date: Tue 11 Jan 2022 09:45:26 PM UTC
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt --pip-args='--no-cache-dir' requirements.in
#
amqp==5.0.9
    # via kombu
asgiref==3.4.1
    # via django
backports.zoneinfo==0.2.1
    # via django
billiard==3.6.4.0
    # via celery
celery==5.2.3
    # via -r requirements.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.10
    # via requests
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
click==8.0.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
django-environ==0.8.1
    # via -r requirements.in
django==4.0.1
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements.in
gevent==21.12.0
    # via gunicorn
greenlet==1.1.2
    # via gevent
gunicorn[gevent]==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
kombu==5.2.3
    # via celery
prompt-toolkit==3.0.24
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements.in
pytz==2021.3
    # via
    #   celery
    #   djangorestframework
requests==2.27.1
    # via -r requirements.in
six==1.16.0
    # via click-repl
sqlparse==0.4.2
    # via django
urllib3==1.26.8
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
zope.event==4.5.0
    # via gevent
zope.interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
